Fix incorrect suffixes for NNTBS retention
Libmtev retention periods for second-scale time maxes out at weeks,
due to the difficulty with years of differing lengths (leap years,
etc.) Use values that will actually work if copied verbatim.

Reword the NNTBS section a bit to be clearer, and move the link to
libmtev durations to the top of the document, since it does not just
concern raw_db.