Additional rebalance details (#68)
Delete after rebalance is not implemented yet.

Abort procedure has changed.

Note that new topology is needed on all nodes, even ones that are
leaving, in the shrinking case.

A config setting is available if the new node(s) cannot handle
simultaneous loading of data from all existing nodes. It is only
checked at the start of a rebalance, not after it has started.